+++ codex-rs/core/src/lib.rs
@@ -28,13 +28,13 @@ pub mod exec_env;
 mod flags;
 pub mod git_info;
 mod is_safe_command;
-mod ishell;
 pub mod landlock;
 mod mcp_connection_manager;
 mod mcp_tool_call;
 mod message_history;
 mod model_provider_info;
 pub mod parse_command;
+mod unified_exec;
 mod user_instructions;
 pub use model_provider_info::BUILT_IN_OSS_MODEL_PROVIDER_ID;
 pub use model_provider_info::ModelProviderInfo;
